requirements.txt
slackclient
googlemaps
pyyaml
